+++ core-ng/src/main/java/core/framework/impl/cache/LocalCacheStore.java
@@ -47,8 +47,10 @@ public class LocalCacheStore implements CacheStore {
     }
 
     @Override
-    public void delete(String key) {
-        caches.remove(key);
+    public void delete(String... keys) {
+        for (String key : keys) {
+            caches.remove(key);
+        }
     }
 
     public void cleanup() {